+++ src/irvine/oeis/a067/A067000.java
@@ -14,7 +14,7 @@ public class A067000 extends CachedSequence {
 
   /** Construct the sequence. */
   public A067000() {
-    super(1, Integer.class, (self, k) -> k == 1
+    super(1, Long.class, (self, k) -> k == 1
       ? Z.ONE
       : Integers.SINGLETON.sum(1, k - 1, j -> self.a(j).multiply(Functions.FACTORIAL.z(k + j)).divide(Functions.FACTORIAL.z(2 * j))));
   }
