+++ src/irvine/oeis/a070/A070864.java
@@ -11,6 +11,6 @@ public class A070864 extends CachedSequence {
 
   /** Construct the sequence. */
   public A070864() {
-    super(1, Integer.class, (self, n) -> n <= 2 ? Z.ONE : Z.TWO.add(self.a(n - self.a(n - 1).intValueExact())));
+    super(1, Long.class, (self, n) -> n <= 2 ? Z.ONE : Z.TWO.add(self.a(n - self.a(n - 1).intValueExact())));
   }
 }
