+++ src/irvine/oeis/a070/A070868.java
@@ -11,6 +11,6 @@ public class A070868 extends CachedSequence {
 
   /** Construct the sequence. */
   public A070868() {
-    super(1, Integer.class, (self, n) -> n <= 2 ? Z.ONE : self.a(self.a(n - 1).intValueExact()).add(self.a(n + 1 - self.a(n - 1).multiply2().intValueExact())));
+    super(1, Long.class, (self, n) -> n <= 2 ? Z.ONE : self.a(self.a(n - 1).intValueExact()).add(self.a(n + 1 - self.a(n - 1).multiply2().intValueExact())));
   }
 }
