+++ src/irvine/oeis/a072/A072491.java
@@ -12,7 +12,7 @@ public class A072491 extends CachedSequence {
 
   /** Construct the sequence. */
   public A072491() {
-    super(0, Integer.class, (self, k) -> k < 4 ? (k > 0 ? Z.ONE : Z.ZERO) : self.a(k - Functions.PREV_PRIME.i(k + 1)).add(1));
+    super(0, Long.class, (self, k) -> k < 4 ? (k > 0 ? Z.ONE : Z.ZERO) : self.a(k - Functions.PREV_PRIME.i(k + 1)).add(1));
   }
 }
