+++ src/irvine/oeis/a072/A072810.java
@@ -13,7 +13,7 @@ public class A072810 extends CachedSequence {
 
   /** Construct the sequence. */
   public A072810() {
-    super(0, Integer.class, (self, n) -> {
+    super(0, Long.class, (self, n) -> {
       if (n == 0) {
         return Z.ONE;
       }
