+++ src/irvine/oeis/a075/A075496.java
@@ -12,7 +12,7 @@ public class A075496 extends CachedSequence {
 
   /** Construct the sequence. */
   public A075496() {
-    super(1, Integer.class, (self, n) -> n == 1 ? Z.ONE : Integers.SINGLETON.sum(1, n - 1, k -> self.a(k).max(self.a(n - k))));
+    super(1, Long.class, (self, n) -> n == 1 ? Z.ONE : Integers.SINGLETON.sum(1, n - 1, k -> self.a(k).max(self.a(n - k))));
   }
 }
