+++ src/irvine/oeis/a076/A076896.java
@@ -11,6 +11,6 @@ public class A076896 extends CachedSequence {
 
   /** Construct the sequence. */
   public A076896() {
-    super(1, Integer.class, (self, n) -> n == 1 ? Z.ONE : Z.valueOf(n).subtract(self.a(2 * n / 3)));
+    super(1, Long.class, (self, n) -> n == 1 ? Z.ONE : Z.valueOf(n).subtract(self.a(2 * n / 3)));
   }
 }
