+++ src/irvine/oeis/a081/A081831.java
@@ -11,6 +11,6 @@ public class A081831 extends CachedSequence {
 
   /** Construct the sequence. */
   public A081831() {
-    super(1, Integer.class, (self, n) -> n <= 2 ? Z.ONE : self.a(self.a(n - self.a(n - 1).intValueExact())).add(self.a(n - self.a(n - 2).intValueExact())));
+    super(1, Long.class, (self, n) -> n <= 2 ? Z.ONE : self.a(self.a(n - self.a(n - 1).intValueExact())).add(self.a(n - self.a(n - 2).intValueExact())));
   }
 }
