+++ src/irvine/oeis/a084/A084002.java
@@ -11,12 +11,12 @@ public class A084002 extends CachedSequence {
 
   /** Construct the sequence. */
   public A084002() {
-    super(1, Integer.class, (self, n) -> {
+    super(1, Long.class, (self, n) -> {
       if (n == 1) {
         return Z.NINE;
       }
       final StringBuilder sb = new StringBuilder();
-      for (int k = n - 1; k > 0; --k) {
+      for (long k = n - 1; k > 0; --k) {
         sb.append(self.a(k));
       }
       for (int k = 2; k < n; ++k) {
